+++ src/main/groovy/life/qbic/datamodel/dtos/business/ProductId.groovy
@@ -12,7 +12,7 @@ import groovy.transform.EqualsAndHashCode
  *
  */
 @EqualsAndHashCode
-class ProductId {
+class ProductId implements Comparable<ProductId>{
 
     /**
      * The type of the identifier is defined by the implementing identifier
@@ -123,5 +123,13 @@ class ProductId {
     @Override
     String toString() {
         return "${type}_${uniqueId}"
+    }
+
+    @Override
+    int compareTo(ProductId productId2) {
+        //if the types are equal compare by uniqueId
+        if(this.equals(productId2)) return 0
+        if(this.type.equals(productId2.type)) return this.uniqueId.compareTo(productId2.uniqueId)
+        return this.type.compareTo(productId2.type)
     }
 }